cloud_functions/requirements.txt
functions-framework==3.*
flask==3.*
line-bot-sdk==3.*
google-generativeai>=0.8.3,<1.0.0
notion-client>=2.0.0,<3.0.0
requests>=2.28.0,<3.0.0
PyYAML>=6.0,<7.0
asgiref>=3.7.0,<4.0.0
google-cloud-firestore>=2.11.0,<3.0.0
